% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm on a Pre-built Graph}
\usage{
dijkstra(graph, start_node, inf = 1e+15)
}
\arguments{
\item{graph}{a list representing the graph, returned by \code{build_graph}.}

\item{start_node}{an integer specifying the source node.}

\item{inf}{a numeric value representing infinity. Defaults to \code{1e15}.}
}
\value{
\code{dijkstra()} returns a numeric vector of length \code{graph$n} containing the shortest distances.
}
\description{
Computes the shortest paths from a single source node to all other nodes
using a graph constructed by \code{build_graph}.
}
\examples{
n <- 4
u <- c(1, 1, 2, 3)
v <- c(2, 3, 4, 4)
w <- c(2, 5, 1, 3)
graph <- build_graph(n, u, v, w)

cost_from_1 <- dijkstra(graph, start_node = 1)
cost_from_4 <- dijkstra(graph, start_node = 4)
}
