% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somniloquy-package.R
\docType{package}
\name{somniloquy-package}
\alias{somniloquy}
\alias{somniloquy-package}
\title{somniloquy: A Collection of Miscellaneous R Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is a personal collection of miscellaneous R functions that were developed during several study groups and personal projects. The package serves as a personal toolkit and a record of learning experiences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryo-asashi/somniloquy}
  \item \url{https://ryo-asashi.github.io/somniloquy/}
  \item Report bugs at \url{https://github.com/ryo-asashi/somniloquy/issues}
}

}
\author{
\strong{Maintainer}: Ryoichi Asashiba \email{ryoichi.asashiba@gmail.com}

}
\keyword{internal}
